add keyborad opt.
